v1.0.0 release

+++ app/core/image.py
@@ -1,19 +1,16 @@
 import os
-import json
 from datetime import datetime
 import shutil
-from dotenv import load_dotenv
 from bs4 import BeautifulSoup
 
-load_dotenv(verbose=True)
 class Image():
     def upload_temp(self,file,type,id=None):
         """
         텍스트 에디터 작성중 임시로 생성되는 사진을 저장
         
         - file : 원하는 사진 파일
-        - type : 1-공지글 2-게시판 3-문제글 4-프로필
-        - id : type이 게시판, 문제글, 프로필일 경우 관련된 id
+        - type : 1-공지글 2-게시판 3-문제글 4-프로필 5-로드맵
+        - id : type이 게시판, 문제글, 프로필, 로드맵일 경우 관련된 id
         """
         nowtime=datetime.now().strftime('%Y%m%d%H%M%S%f')
         extension=file.filename.split(".")[-1].lower()
@@ -32,6 +29,9 @@ class Image():
         elif type==4:#프로필 이미지
             filepath=os.getenv("PROFILE_PATH")
             filename=id+".jpg"
+        elif type==5:#로드맵 이미지
+            filepath=os.path.join(os.getenv("ROADMAP_PATH"),"temp",id)
+            filename=nowtime+"."+extension
 
         if not os.path.exists(filepath):
             os.makedirs(filepath)
@@ -45,8 +45,8 @@ class Image():
         저장된 사진의 경로 출력
         
         - filename : 원하는 사진 파일
-        - type : 1-공지글 2-게시판 3-문제글 4-프로필
-        - id : type이 게시판, 문제글, 프로필일 경우 관련된 id
+        - type : 1-공지글 2-게시판 3-문제글 4-프로필 5-로드맵
+        - id : type이 게시판, 문제글, 프로필, 로드맵일 경우 관련된 id
         """
         print("dd")
         filepath=""
@@ -60,7 +60,9 @@ class Image():
             filepath=os.getenv("PROFILE_PATH")
             if not os.path.exists(os.path.join(filepath,filename)):
                 shutil.copyfile(os.path.join(filepath,"base.jpg"),os.path.join(filepath,filename))
-                
+        elif type==5:#로드맵 이미지
+            filepath=os.path.join(os.getenv("ROADMAP_PATH"),str(id))
+
         target_path=os.path.join(filepath,filename)
         if not os.path.exists(target_path):
             return None
@@ -71,8 +73,8 @@ class Image():
         임시로 저장된 사진의 경로 출력
         
         - filename : 원하는 사진 파일
-        - type : 1-공지글 2-게시판 3-문제글 4-프로필
-        - id : type이 게시판일 경우 작성한 userid
+        - type : 1-공지글 2-게시판 3-문제글 4-프로필 5-로드맵
+        - id : type이 게시판,로드맵일 경우 작성한 userid
         """
         filepath=""
         if type==1:#공지글 이미지
@@ -83,13 +85,15 @@ class Image():
             filepath=os.path.join(os.getenv("TASK_PATH"),"temp")
         elif type==4:#프로필 이미지
             filepath=os.getenv("PROFILE_PATH")
+        elif type==5:#로드맵 이미지
+            filepath=os.path.join(os.getenv("ROADMAP_PATH"),"temp",id)
 
         target_path=os.path.join(filepath,filename)
         if not os.path.exists(target_path):
             return None
         return target_path
 
-    def save_update_image(self,temp_path:str,save_path:str,data:str,id:int=None,mode:str="su"):
+    def save_update_image(self,temp_path:str,save_path:str,data:str,id:int|str=None,mode:str="su"):
         """
         data에서 실제 저장되는 사진들만 temp_path에서 save_path로 이동하고 나머지 temp_path는 삭제하거나
         data가 수정되어 save_path의 사진들중 필요없는 사진을 삭제
@@ -111,11 +115,14 @@ class Image():
         img_tags=soup.find_all('img')
         # 저장된 data에서 쓰인 사진만 추출 및 텍스트 에디터의 사진 경로를 실제 사진 경로로 수정
         for i in range(len(img_tags)):
-            image_url=img_tags[i]["src"].split("/temp")[0]
+            image_url=img_tags[i]["src"].split("/temp")[0].split("?id")[0]
             img_tags[i]["src"]=image_url
             imagelist.append(image_url.split("/")[-1])
             if id:
-                img_tags[i]["src"]+="?id="+str(id)
+                try:
+                    id=str(id)
+                finally:
+                    img_tags[i]["src"]+="?id="+id
         new_data=str(soup)
 
         if "s" in mode:
@@ -137,7 +144,10 @@ class Image():
             if os.path.exists(save_path):
                 save_imagelist=os.listdir(save_path)
                 for i in save_imagelist:
-                    if not i in imagelist:
+                    temp=i.split(".")
+                    if len(temp)<=1:
+                        continue
+                    if temp[-1]!="zip" and not i in imagelist:
                         os.remove(os.path.join(save_path,i))
 
         return new_data
